+++ src/parse/parse-direcetory.ts
@@ -1,4 +1,4 @@
-import { none, Some, Maybe } from '../maybe';
+import { Maybe } from '@nkp/maybe';
 import { InputChild } from './parse-type';
 
 export type InputDirectoryArray = [ name: string, children: InputChild[] ];
@@ -8,23 +8,23 @@ export type InputDirectory = InputDirectoryArray | InputDirectoryObject;
 export type InputNormalizedDirectory = { type: 'directory', name: string, children: unknown[] };
 
 export function parseDirectory(unknown: unknown): Maybe<InputNormalizedDirectory> {
-  if (!unknown) return none;
+  if (!unknown) return Maybe.none;
   const _unknown = unknown as InputDirectory;
 
   if (Array.isArray(_unknown)) {
-    if (_unknown.length !== 2) return none;
+    if (_unknown.length !== 2) return Maybe.none;
     const [name, children,] = _unknown;
-    if (!Array.isArray(children)) return none;
-    if (typeof name !== 'string') return none;
-    return new Some({ type: 'directory', name, children, });
+    if (!Array.isArray(children)) return Maybe.none;
+    if (typeof name !== 'string') return Maybe.none;
+    return Maybe.some({ type: 'directory', name, children, });
   }
 
   else if (typeof _unknown === 'object') {
     const { name, children, } = _unknown;
-    if (typeof name !== 'string') return none;
-    if (!Array.isArray(children)) return none;
-    return new Some({ type: 'directory', name, children, });
+    if (typeof name !== 'string') return Maybe.none;
+    if (!Array.isArray(children)) return Maybe.none;
+    return Maybe.some({ type: 'directory', name, children, });
   }
 
-  return none;
+  return Maybe.none;
 }