+++ src/parse/parse-file.ts
@@ -1,7 +1,7 @@
-import { none, Some, Maybe } from '../maybe';
 import { OrPromise } from '../types';
 import { FileWrite } from '../nodes/file';
 import { InputType } from './parse-type';
+import { Maybe } from '@nkp/maybe';
 
 export type InputFileWriteObject = { content: string | Buffer, encoding?: BufferEncoding };
 export type InputFileWriteFn = () => OrPromise<InputFileWriteObject>;
@@ -14,12 +14,12 @@ export type InputFile = InputFileArray | InputFileObject;
 export type InputNormalizedFile = { type: InputType.File, name: string, write: FileWrite }
 
 export function parseFile(unknown: unknown): Maybe<InputNormalizedFile> {
-  if (!unknown) return none;
+  if (!unknown) return Maybe.none;
   const _unknown = unknown as InputFile;
   if (Array.isArray(_unknown)) {
-    if (_unknown.length !== 2) return none;
+    if (_unknown.length !== 2) return Maybe.none;
     const [name, options,] = _unknown;
-    if (typeof name !== 'string') return none;
+    if (typeof name !== 'string') return Maybe.none;
     return parseFileWriteFn(options).map((write): InputNormalizedFile => ({
       type: InputType.File,
       name,
@@ -28,24 +28,24 @@ export function parseFile(unknown: unknown): Maybe<InputNormalizedFile> {
   }
   else if (typeof _unknown === 'object') {
     const { name, write, } = _unknown;
-    if (typeof name !== 'string') return none;
+    if (typeof name !== 'string') return Maybe.none;
     return parseFileWriteFn(write).map((write): InputNormalizedFile => ({
       type: InputType.File,
       name,
       write,
     }));
   }
-  return none;
+  return Maybe.none;
 }
 
 function parseFileWriteFn(options: unknown): Maybe<FileWrite> {
-  if (!options) return none;
+  if (!options) return Maybe.none;
   switch (typeof options) {
   case 'string':
-    return new Some(() => ({ content: options, encoding: 'utf-8', }));
+    return Maybe.some(() => ({ content: options, encoding: 'utf-8', }));
 
   case 'function':
-    return new Some(async () => {
+    return Maybe.some(async () => {
       const { content, encoding, } = await (options as InputFileWriteFn)();
       return { content, encoding: encoding ?? 'utf-8', };
     });
@@ -55,10 +55,10 @@ function parseFileWriteFn(options: unknown): Maybe<FileWrite> {
 
     switch (typeof write) {
     case 'string':
-      return new Some(() => ({ content: write, encoding: _encoding ?? 'utf-8', }));
+      return Maybe.some(() => ({ content: write, encoding: _encoding ?? 'utf-8', }));
 
     case 'function': {
-      return new Some(async () => {
+      return Maybe.some(async () => {
         const { content, encoding, } = await write();
         return { content, encoding: encoding ?? _encoding ?? 'utf-8', };
       });
@@ -67,5 +67,5 @@ function parseFileWriteFn(options: unknown): Maybe<FileWrite> {
   }
   }
 
-  return none;
+  return Maybe.none;
 }