+++ src/parse/parse-symlink.ts
@@ -1,4 +1,4 @@
-import { none, Some, Maybe } from '../maybe';
+import { Maybe } from '@nkp/maybe';
 import { InputType } from './parse-type';
 
 export type InputSymlinkOptions = { to: string }
@@ -9,25 +9,25 @@ export type InputSymlink = InputSymlinkObject | InputSymlinkArray;
 export type InputNormalizedSymlink = { type: InputType.File, name: string, link: string, }
 
 export function parseSymlink(unknown: unknown): Maybe<InputNormalizedSymlink> {
-  if (!unknown) return none;
+  if (!unknown) return Maybe.none;
   const _unknown = unknown as InputSymlink;
   if (Array.isArray(_unknown)) {
-    if (_unknown.length !== 2) return none;
+    if (_unknown.length !== 2) return Maybe.none;
     const [name, options,] = _unknown;
-    if (typeof name !== 'string') return none;
-    if (!options || typeof options !== 'object') return none;
+    if (typeof name !== 'string') return Maybe.none;
+    if (!options || typeof options !== 'object') return Maybe.none;
     const { type, link, } = options;
-    if (type !== undefined && type !== InputType.Symlink) return none;
-    if (typeof link !== 'string') return none;
-    return new Some({ type: InputType.File, name, link, });
+    if (type !== undefined && type !== InputType.Symlink) return Maybe.none;
+    if (typeof link !== 'string') return Maybe.none;
+    return Maybe.some({ type: InputType.File, name, link, });
   }
   if (typeof _unknown === 'object') {
     const { name, type, link, } = _unknown;
-    if (typeof name !== 'string') return none;
-    if (type !== undefined && type !== InputType.Symlink) return none;
-    if (typeof link !== 'string') return none;
-    return new Some({ type: InputType.File, name, link, });
+    if (typeof name !== 'string') return Maybe.none;
+    if (type !== undefined && type !== InputType.Symlink) return Maybe.none;
+    if (typeof link !== 'string') return Maybe.none;
+    return Maybe.some({ type: InputType.File, name, link, });
   }
-  return none;
+  return Maybe.none;
 }